# docs/swagger.yaml
definitions:
  schema.GameSearchRequestDto:
    properties:
      category:
        items:
          type: integer
        type: array
      genres:
        items:
          type: string
        type: array
      limit:
        type: integer
      page:
        type: integer
      platforms:
        items:
          type: string
        type: array
      profile:
        type: boolean
      query:
        type: string
      status:
        items:
          type: integer
        type: array
      themes:
        items:
          type: string
        type: array
    type: object
  schema.GameSearchResponseDto:
    properties:
      data:
        $ref: '#/definitions/schema.ResponseData'
      pagination:
        $ref: '#/definitions/schema.PaginationInfo'
    type: object
  schema.PaginationInfo:
    properties:
      hasNextPage:
        type: boolean
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
  schema.ResponseData:
    properties:
      items:
        items:
          $ref: '#/definitions/schema.SearchGame'
        type: array
      profile:
        additionalProperties: true
        type: object
      took:
        type: integer
    type: object
  schema.SearchGame:
    properties:
      aggregatedRating:
        type: number
      aggregatedRatingCount:
        type: integer
      category:
        type: integer
      checksum:
        type: string
      coverUrl:
        type: string
      createdAt:
        type: string
      firstReleaseDate:
        type: string
      genresNames:
        type: string
      id:
        type: integer
      keywordsNames:
        type: string
      name:
        type: string
      platformsAbbreviations:
        type: string
      platformsNames:
        type: string
      slug:
        type: string
      source:
        type: string
      status:
        type: integer
      storyline:
        type: string
      summary:
        type: string
      themesNames:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact:
    email: support@gamenode.com.br
    name: API Support
    url: gamenode.com.br/help
  description: The GameNode Search API documentation.
  license:
    name: GNU General Public License
    url: http://www.gnu.org/licenses/
  termsOfService: http://swagger.io/terms/
  title: GameNode Search API
  version: "1.0"
paths:
  /search/games:
    post:
      consumes:
      - application/json
      description: Returns a parsed search response from the Manticore engine
      parameters:
      - description: Account ID
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/schema.GameSearchRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.GameSearchResponseDto'
      summary: Searches for games using Manticore engine
      tags:
      - search
swagger: "2.0"